LCORE-2850: Renamed pydantic_ai utility module
